Add support for merging metadata when merging LogEvent's
This moves the merge_log_event and merge_value functions into the impl
for LogEvent and Value.
