+++ src/event/merge_state.rs
@@ -1,4 +1,3 @@
-use crate::event::merge::merge_log_event;
 use crate::event::LogEvent;
 
 /// Encapsulates the inductive events merging algorithm.
@@ -22,7 +21,7 @@ impl LogEventMergeState {
 
     /// Merge the incoming (partial) event in.
     pub fn merge_in_next_event(&mut self, incoming: LogEvent, fields: &[impl AsRef<str>]) {
-        merge_log_event(&mut self.intermediate_merged_event, incoming, fields);
+        self.intermediate_merged_event.merge(incoming, fields);
     }
 
     /// Merge the final (non-partial) event in and return the resulting (merged)
