+++ src/event/value.rs
@@ -1,5 +1,5 @@
 use crate::{event::timestamp_to_string, Result};
-use bytes::Bytes;
+use bytes::{Bytes, BytesMut};
 use chrono::{DateTime, Utc};
 use derive_is_enum_variant::is_enum_variant;
 use serde::{Serialize, Serializer};
@@ -288,6 +288,21 @@ impl Value {
             Value::Map(_) => "map",
             Value::Array(_) => "array",
             Value::Null => "null",
+        }
+    }
+
+    /// Merges `incoming` value into self.
+    ///
+    /// Will concatenate `Bytes` and overwrite the rest value kinds.
+    pub fn merge(&mut self, incoming: Value) {
+        match (self, incoming) {
+            (Value::Bytes(self_bytes), Value::Bytes(ref incoming)) => {
+                let mut bytes = BytesMut::with_capacity(self_bytes.len() + incoming.len());
+                bytes.extend_from_slice(&self_bytes[..]);
+                bytes.extend_from_slice(&incoming[..]);
+                *self_bytes = bytes.freeze();
+            }
+            (current, incoming) => *current = incoming,
         }
     }
 }
